add cdb fixture generation so tests can generate their own CDB files exactly as needed.
The current fixture generation is somewhat basic. For example it still relies
on real (temporary) files while ideally all testing should be done completely
in-memory (less side effects, no need for tearDown methods, faster) but that
requires some more code to emulate a mmap'ed file based on a bytes instance.

Also the IBF file name can not be set but from my point of that are minor
details which will be fixed when the fixture generation is used to write
"real" tests (e.g. for form validation).