remove (test) dependency on pillow
tiffany can not create tiff images and we should try not to add new
dependencies. Currently there is no test at all which relies on having actual
tiff data so we can just use some random binary data.
We might have to reconsider this decision once we need to test tiff-related
functionality.

Also the production setup uses irregular tiff files which we likely can't
synthesize (which means we'd have to include static files in the repository).