Bail out when parsing a CDB file with unknown field names (pydica issue 10)
The old software sometimes writes junk for some form fields. That seems to
happen in the old software if a user entered more charcters than the field
definition actually allows. The extra character will overflow in a new
"field". We can catch that by ensuring that we only accept known field
names.