nasty locking problem on windows is solved.
When initializing a batch, the introducer creates a snapshot of the CDB data.
But this collides with the locking of the CDB that happened before.

Short solution was to move the creation of the CDB object behind the snapshot.
Long term solution when we use this:
- the scapshot should re-use the file descriptor and not open the CDB file a second time.
(will do that later, but now it works on Windows)