also use smaller font-size on Linux
This is likely a hack - I guess we should really use QT to detect the
width/height of rendered fonts and adapt the font size so the rendered
glyphs fit into a bounding box (form input).

However it works for me right now with very little effort so let's
go with this hack for now.