create new helper function/constant to get the names of all known fields.
This removes a bit of code duplication and not-so-easy-to-understand code
for something which should be a trivial task. Also the caps indicate that
we assume field definitions will not change after initialization.

Copying some BSD-licensed code from Babel to avoid another dependency
(although there should not be much harm in depending from Babel: It
is pure Python and supports Python 3 for quite some time).