remove old "image_filename()" helpers
Actually FormBatch instances do not need to know the name of the corresponding
image files and just computing them based on a file-name is prone to many
bugs which are solved in our current helper libraries (e.g. case-sensitive
filesystems).

As a nice side effect we can shave off quite a few lines of duplicated code.