rename "pic_nr" property in cdb_tool.Form to "cdb_pic_nr"
The PIC number is stored in different places (CDB form header and
in the tiff header of every page). When a form is marked as deleted
the CDB form header and the tiff header of the first scanned page
are set to "DELETED" while the original PIC number is left in the
tiff header of the second scanned page.

It seems dangerous just to use "Form.pic_nr" because the developer
might not be aware of the intricacies of getting the "right" PIC
number.

Also I would like to avoid open-coding the "is deleted" check in
multiple places (i.e. explicitely comparing the PIC number against
"DELETED"). Instead all code should just access the helper functions
which shield the developer from these storage implementation details.