remote control: prefer non-deleted forms even in OPENDEL
Previously the OPENDEL command always returned the first form for
a given PIC (regardless of its deletion state).

However it turned out that it is much more useful in daily work to
prefer non-deleted forms over deleted ones. Most of the times
forms are deleted already in the scan process (and re-scanned
immediately) so showing the first deleted form is likely not
helpful.

In case there is no non-deleted form for a given PIC OPENDEL will
return the index of the deleted form with the highest index which
is likely the form which was deleted manually.