prevent delayed loading of forms when opening a RDB/CDB file
The main problem is that this is not transparent for the caller when
accessing form data via cdb_tool.FormBatch.forms. If delay_load is
True, that list contains callables instead of actual instances.

This complicates the code quite a bit and I think delaying the loading does
not affect the performance that much so disabling for now.

Might be enabled again if callers don't have to care how the FormBatch
was initialized.