Batch: only store plain paths in bunch
A DataBunch might contain a file-like object (which is very convenient and
gives us more flexibility in how we assemble Batch instances) but other
code sometimes needs to know the actual file paths.

To prevent errors the Batch now already deduces the paths (and stores
None if it could not deduct the path).