implement basic CDB sanity check
The idea is to check if a CDB can be opened and seems to contain valid
data (binary structure, no semantic validation of prescription values)
before we try to do the real parsing.

The parser in FormBatch is not as straight-forward as I'd like and in
general FormBatch does not have a good place to report basic errors
like "unable to open file". Therefore I'm adding a basic sanity checker
which is designed to satisfy also GUI needs (e.g. provide helpful
messages). More checks will be added in later commits.