CDB_Form: use TiffHandler instances from Batch
Previously we had two (potentially) different instances representing the
same TIFF image which means there is the possibility of inconsistencies.
Probably this did not cause any bugs because we just look at the PIC
numbers and only one of the instances was modified/used for deletion
checks.

Still it is for sure beneficial to have just a single representation
at any time.

Also this is another step in getting rid of CDB_Collection.