CDB fields: make valid marker field accessible via cdb_tool.Form
I don't really understand all the magic in WithBinaryMeta which maps
the format specification to Form/FormField classes but it seems that
names starting with "_" are just ignored (they don't appear in
"field.rec...").

With this change we can access the "valid" field which the Walther
software uses to flag a field value as "invalid" (0=invalid, 1=valid).