DataBunch: remove ability to omit parameters
That feature was introduced in git commit 6773d50c. It seems as if this commit
introduced a really nasty bug where arguments for the DataBunch constructor
would not be stored in the correct attributes, essentially assigning the
parameters to (seemingly) random attributes.

I am not sure how this could happen and I was unable to reproduce the issue
with unit tests or on a test system (with Python 3.6) but it seems to happen
quite often with Python 3.4.

This reverts most of commit:

  commit 6773d50cf827e6438ea59fc92e6a833dae821d21
  Author: Felix Schwarz <felix.schwarz@oss.schwarz.eu>
  Date:   Fri Apr 27 12:21:42 2018 +0200

    DataBunch: make all parameters optional