FormBatch: remove check for "unknown fields"
"open_cdb()" does a much better job at checking the CDB structure. If a safe-guard
is needed you should run ""open_cdb()" first and pass the "fp" returned by that
function to FormBatch.

Removing the field names in FormBatch helps callers to reduce the number of
places which need a reference to the expected/known field names (which in
turn can be used to remove unnecessary dependencies).