downgrade TIFF serialization to version from pillow 8.4.0, bump minimum requirement to pillow 8.3.0
We need at least pillow 8.3.0 because of TiffTags.TAGS_V2_GROUPS.
I came to the conclusion that I should not use a really new version (9.x) which
dropped support for Python 3.6. pillow 8.4 is the last version with 3.6 support
and I would like to keep that in case rdblib is used on our servers.
Backporting really new versions always comes with a risk that I introduce
subtle incompatibilities.

Minimum requirement of 8.3.0 is convenient (over 8.4.0) as 8.3.2 is the last version
which contains binary wheels for Linux for Python 3.6.